refactor(ingestion): consolidate duplicate extract_case_title functions (#1405)
The backfill_case_titles.py script had its own extract_case_title() with
5 extraction strategies, while ingestion/extract.py had a separate
implementation. Any new pattern needed to be added to both files.

Merge all strategies (caption block, Case Name field, MOVING/RESPONDING
PARTY, inline v. patterns, In re) into extract.py and have the backfill
script import from there. Remove the duplicate "Case Name:" regex from
_CASE_TITLE_PATTERNS since it is now handled by the more precise
_extract_from_case_name_field (which validates "v." presence).
